feat: Begin working on modular io workers - Move common functionality out of io_uring specific code - Create io_uring specific code path - Begin structuring crate in a way that will make it easier to swap backend
