+++ Cargo.toml
@@ -11,18 +11,14 @@ members = [".", "fuzz", "verified"]
 
 [features]
 default = []
-io-uring = ["dep:io-uring"]
 shuttle = ["dep:shuttle"]
 
 [dependencies]
 libc = "0.2"
+io-uring = "0.7"
 shuttle = { version = "0.8", optional = true }
 verified = { package = "t4-verified", version = "0.1.3", path = "verified" }
 vstd = "=0.0.0-2026-02-22-0103"
-
-[target.'cfg(linux)'.dependencies]
-io-uring = { version = "0.7", optional  = true }
-
 
 [dev-dependencies]
 pollster = "0.4"
